! Typo in function declaration (install_2-0_postgresql.sql) [Bug 3131]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@7948 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/install_2-0_postgresql.sql
@@ -29,7 +29,7 @@ CREATE OR REPLACE FUNCTION IFNULL (character varying, boolean) RETURNS character
   'SELECT COALESCE($1, CAST(CAST($2 AS int) AS varchar)) AS result'
 LANGUAGE 'sql';
 
-CREATE OR REPLACE FUNCTION IFNULL (int, boolean) RETURNS int8 AS
+CREATE OR REPLACE FUNCTION IFNULL (int, boolean) RETURNS int AS
   'SELECT COALESCE($1, CAST($2 AS int)) AS result'
 LANGUAGE 'sql';
